better menu generic component
